Added implementation examples and refactored the new end point classes to be more consistent with previous naming conventions. Corrected a typo in the inquiry model setting the personid as a string when it is an int. Also corrected a comment in the inquiry_insert script.
