+++ examples/inquiry_insert.php
@@ -4,10 +4,10 @@
 chdir(__DIR__);
 require 'init.i.php';
 
-// create a new person and fill in details (name)
+// create a new inquiry and fill in details
 $inquiry = new \Arctic\Model\Inquiry\Inquiry();
 $inquiry->personid = 199;
 $inquiry->notes = str_repeat('test ', 23110);
 
-// insert both the inquiry
+// insert the inquiry
 $inquiry->insert();
