+++ examples/tripformresponse_browse.php
@@ -1,0 +1,13 @@
+<?php
+
+// set directory
+chdir(__DIR__);
+require 'init.i.php';
+
+// run browse
+//  gets the first 5 responses in the database
+//  0   is the starting index
+//  5   is the number of entries to return
+foreach (\Arctic\Model\TripForm\Response::browse(0,5) as $response ) {
+	echo '* ' , $response->id , ': ' , $response->tripid , ' ' , $response->businessgroupid , "\n";
+}
