+++ includes/Model/Evaluation/Response.php
@@ -1,26 +1,26 @@
 <?php
 
-namespace Arctic\Model\TripFormResponse;
+namespace Arctic\Model\Evaluation;
 
 use Arctic\Model;
 
 /**
- * Class TripFormResponse
+ * Class Response
  * @property int $businessgroupid
- * @property int $triptripformid
- * @property int $tripformformid
+ * @property int $evaluationid
  * @property int $id
- * @property int $tripid
  * @property int $personid
- * @property DateTime $time
- * @property \Arctic\Model\Trip\Trip $trip
+ * @property int $activityid
+ * @property string $useragent
+ * @property \DateTime $time
+ * @property \Arctic\Model\Activity\Activity $activity
  * @property \Arctic\Model\BusinessGroup $businessgroup
  * @property \Arctic\Model\Person\Person $person
  */
-class TripFormResponse extends Model
+class Response extends Model
 {
 	public static function getApiPath() {
-		return 'tripformresponse';
+		return 'evaluationresponse';
 	}
 
 	public function __construct() {
@@ -28,6 +28,6 @@ class TripFormResponse extends Model
 
 		$this->_addSingleReference( 'businessgroup' , 'Arctic\Model\BusinessGroup' , array( 'businessgroupid' => 'id' ) );
 		$this->_addSingleReference( 'person' , 'Arctic\Model\Person\Person' , array( 'personid' => 'id' ) );
-		$this->_addSingleReference( 'trip' , 'Arctic\Model\Trip\Trip' , array( 'tripid' => 'id' ) );
+		$this->_addSingleReference( 'activity' , 'Arctic\Model\Activity\Activity' , array( 'activityid' => 'id' ) );
 	}
 }
